+++ bin/pato-git-workflow.sh
@@ -25,9 +25,6 @@ These are the options:
 - merge (from/to a branch)
 - merge_abort
 - release (from lowest to highest protected branch)
-- release_backup (from lowest to highest protected branch)
-- release_copy (from lowest to highest protected branch)
-- release_install (from lowest to highest protected branch)
 
 
 info
@@ -72,29 +69,39 @@ Issues "git merge --abort".
 
 
 release <protected_branch_1> <protected_branch_2> ... <protected_branch_N>
-release_backup <protected_branch_1> <protected_branch_2> ... <protected_branch_N>
-release_copy <protected_branch_1> <protected_branch_2> ... <protected_branch_N>
-release_install <protected_branch_1> <protected_branch_2> ... <protected_branch_N>
 --------------------------------------------------------------------------
 All branches must be protected.
-The branch <protected_branch_1> will usually be 'development', <protected_branch_2> 'acceptance' and <protected_branch_3> 'main' or 'master' (N = 3).
+
+The branch <protected_branch_1> will usually be 'development',
+<protected_branch_2> 'acceptance' and
+ <protected_branch_3> 'main' or 'master' (N = 3).
+
 This is a prompted workflow, meaning that at every step the user is asked to do something.
 
+While executing the release command steps, a release state file is updated
+so you restart from where you left in case of an error (or quitting from your side).
+This file is located in the \`target\` folder and starts with \`pato-git-workflow\`.
+
 These are the steps:
-0. Each branch <protected_branch_n> will be copied to export/<protected_branch_n> and the user will be asked to create an export using the PATO GUI (export APEX and generate DDL). This is a safety measure that is MANDATORY.
-1. Next, branch <protected_branch_1> may need to be installed first to ensure that all development stuff is correctly installed from feature branches.
-   A user action (with the PATO GUI) for APEX and/or database is needed in ANOTHER SESSION.
+0. Each branch <protected_branch_n> will be copied to export/<protected_branch_n> and
+   the user will be asked to create an export using the PATO GUI (export APEX and generate DDL).
+   This is a safety measure that is MANDATORY.
+1. Next, branch <protected_branch_1> may need to be installed first to ensure that all
+   development stuff is correctly installed from feature branches.
+   A database install (with the PATO GUI) is needed in ANOTHER SESSION.
 2. Next every branch n will be released to release/<protected_branch_n>-<protected_branch_m> (m = n+1):
-   a. using the copy command first, see above (no user action needed)
-   b. export APEX and/or database (generate DDL) next (for environment n) using the PATO GUI (user action in ANOTHER SESSION)
+   a. using the copy command first, see above (no user action needed):
+      the first release branch will copy from the modified <protected_branch_1> (due to the database install in step 1),
+      the rest from export/<protected_branch_n> (n > 1)
+   b. export APEX and/or database (generate DDL) next (for branch/environment <protected_branch_1> only)
+      using the PATO GUI (user action in ANOTHER SESSION)
    c. pushing the changes back to the GitHub repo (no user action needed)
 3. Next every release branch (release/<protected_branch_n>-<protected_branch_m>) is merged into <protected_branch_m> (m = n+1):
-   a. use a Pull Request
-   b. tag the resulting branch with "\`basename \$0 .sh\`-\`date -I\`" (for example pato-git-workflow-2025-03-04)
-4. Now every protected branch has the new code, so just install the branches m, 1 < m <= N. Please note that branch 1 has already been installed.
+   - by using the merge strategy 'ours', effectively replacing branch <protected_branch_m> (but keeping history)
+   - the resulting branch will be tagged with "\`basename \$0 .sh\`-\`date -I\`" (for example pato-git-workflow-2025-03-04)
+4. Now every protected branch has the new code, so just install the branches m, 1 < m <= N.
+   Please note that branch 1 has already been installed.
 
-Option release invokes release_backup (step 0) and release_copy (steps 1 till 3) to prepare the release and then release_install to install it (step 4).
-This allows for initializing, copying and installing a release on different moments.
 
 ENVIRONMENT VARIABLES
 =====================
@@ -159,9 +166,17 @@ Since you do not like to type in more than needed, you can make an alias in your
 
 Now you can just type pgw.
 
-The export from development and merging it back into development from above can now be written as (provided variable r is a repo and you are in folder ~/dev):
+The export from development and merging it back into development from above
+can now be written as (provided variable r is a repo and you are in folder ~/dev):
 
-  (set -e; cd \$r; b=development; pgw copy \$b export/\$b; cd ../pato-gui; pato-gui ../\$r/db/pom.xml; pato-gui ../\$r/apex/pom.xml; cd ../\$r; pgw merge export/\$b \$b -X ours)
+  set -e
+  cd \$r; b=development
+  pgw copy \$b export/\$b
+  cd ../pato-gui
+  pato-gui ../\$r/db/pom.xml
+  pato-gui ../\$r/apex/pom.xml
+  cd ../\$r
+  pgw merge export/\$b \$b -X ours
 
 alias pato-gui:
 
@@ -208,7 +223,7 @@ _ignore_output() { eval "$@" $ignore_output; }
 
 export -f _ignore_stdout _ignore_stderr _ignore_output
 
-test -n "${PROTECTED_BRANCHES:-}" || PROTECTED_BRANCHES="development test acceptance main master"
+[[ -n "${PROTECTED_BRANCHES:-}" ]] || PROTECTED_BRANCHES="development test acceptance main master"
 export PROTECTED_BRANCHES
 
 _prompt() {
@@ -272,7 +287,7 @@ _check_upstream_exists() {
 }
 
 _check_no_changes() {
-    test -z "$(git status --porcelain)" || { _error "You have changes in your workspace"; git status; exit 1; }
+    [[ -z "$(git status --porcelain)" ]] || { _error "You have changes in your workspace"; git status; exit 1; }
 }
 
 _switch() {
@@ -283,7 +298,8 @@ _switch() {
 }
 
 _tag() {
-    declare -r tag="`basename $0 .sh`-`date -I`"
+    declare -r branch=${1:?}
+    declare -r tag="`basename $0 .sh`-`date -I`-${branch}"
     
     _x git tag $tag
 }
@@ -294,8 +310,12 @@ _pull_request() {
 
     if `git remote -v | grep github 1>/dev/null`
     then
-        _x gh pr create --title "$from => $to" --editor --head $from --base $to
-        _x gh pr view --web
+        if _x gh pr create --title "$from => $to" --editor --head $from --base $to
+        then
+            _x gh pr view --web $from
+        else
+            _prompt "Something went wrong"
+        fi
     elif `git remote -v | grep azure 1>/dev/null`
     then
         _x az repos pr create \
@@ -338,13 +358,13 @@ copy() {
     _check_branch_not_protected $to || _error_branch_protected $to
     _check_no_changes
     _switch $from
-    _x git pull
+    # _x git pull
     
     # $from exists but $to maybe not
     if ! _ignore_stderr git checkout -b $to $from
     then
         # delete local and remote branch (just pointers)
-        _prompt "Remove local (and remote) branch $to before re-creating it from branch $from"
+        _prompt "This script is about to remove local (and remote) branch $to before re-creating it from branch $from"
         _x git branch -D $to
         _x git push origin --delete $to || true
         _x git checkout -b $to $from
@@ -352,10 +372,11 @@ copy() {
     _x git push --set-upstream origin $to
     if [[ -n "${DEBUG:-}" ]]
     then
-        _prompt "Showing \"git diff $from $to\" for both local and remote branches"
+        _prompt "This script is about to show \"git diff $from $to\" for both local and remote branches"
         _x git diff --name-status $from $to
         _x git diff --name-status origin/$from origin/$to
     fi
+    _info "Copied branch $from to $to"
 }
 
 merge() {
@@ -371,7 +392,7 @@ merge() {
 
     # merge the changes made to $to in the meantime back into $from before we will merge back
     _switch $from
-    _x git merge $to || _prompt "Fix the conflicts"
+    _x git merge $to || _prompt "You must fix the conflicts"
     _x git commit -m "Make $from up to date with $to" || true
 
     # now the real merge
@@ -381,7 +402,7 @@ merge() {
         _x git merge $options $from
     else
         # to merge into a protected branch we need a Pull Request
-        _prompt "Pushing branch $from"
+        _prompt "This script is about to push branch $from"
         _x git push # push $from to remote
         _switch $to # must be on a branch named differently than "release/acceptance-main"
         _pull_request $from $to
@@ -392,16 +413,68 @@ merge_abort() {
     _x git merge --abort
 }
 
-release_backup() {
+declare release_state_file=
+_release_state_file_reuse() { # usage: PROTECTED_BRANCHE_1 ... PROTECTED_BRANCHE_N
+    declare yes_no=
+
+    release_state_file="target/$(basename $0 .sh)-release-$(echo $* | sed -e 's/ /-/g').txt"
+
+    if [[ -f ${release_state_file} ]]
+    then
+        read -p "Do you want to reuse the previous state (from file ${release_state_file}) ? [Y] " yes_no
+        case "${yes_no}" in
+            n | N)
+                rm ${release_state_file}
+                ;;
+            *)
+                :
+                ;;
+        esac        
+    fi
+}
+
+_release_state_file_cleanup() {
+    test ! -f ${release_state_file} || mv ${release_state_file} "${release_state_file}~"
+}
+
+_release_step_skip() { # usage: BRANCH STEP
+    declare -r branch=${1:?}
+    declare -r step=${2:?}
+
+    # do not show the result
+    [[ -f ${release_state_file} ]] && grep -E "^${branch}:${step}$" ${release_state_file} 1>/dev/null
+
+    declare -r result=$? # 0 it exists
+
+    [[ $result -ne 0 ]] || _info "Skipping step ${step} for branch ${branch} since it has already been executed"
+    
+    return $result
+}
+
+_release_step_write() { # usage: BRANCH STEP
+    declare -r branch=${1:?}
+    declare -r step=${2:?}
+
+    [[ -d target ]] || mkdir target
+    echo "${branch}:${step}" >> ${release_state_file}
+    _info "Finished step ${step} for branch ${branch}"
+}
+
+release() {
     declare -r branches=$*
     declare from=
     declare to=
-    declare release=
+    declare first=
+    declare step=
     declare export=
+    declare release=
 
     _check_no_changes
 
+    _release_state_file_reuse $*
+
     # step 0 from usage for release
+    step=0
     for to in $branches
     do
         _check_upstream_exists $to || _error_upstream_does_not_exist $to
@@ -416,7 +489,6 @@ release_backup() {
                     "test|acceptance" | \
                     "acceptance|main" | \
                     "acceptance|master")
-                    echo "Releasing from $from to $to"
                     ;;
                 *)
                     _error "Wrong combination for release: $from => $to"
@@ -424,90 +496,131 @@ release_backup() {
                     ;;
             esac
         fi
+
+        if ! _release_step_skip $to $step
+        then
+            export="export/$to"
             
-        export="export/$to"
-        copy $to $export
-        _prompt "Please create a backup export (APEX and/or database) in ANOTHER SESSION (using PATO GUI with POMs from $pwd/apex and $pwd/db)"
-        _prompt "Pushing branch $export"
-        _x git push
+            copy $to $export
+            _prompt "You must create an export (APEX and/or database) for '$to' in ANOTHER SESSION (using PATO GUI with POMs from $pwd/apex and $pwd/db)"
+            _x git add .
+            _x git commit -m "Created export $export" || true
+            _prompt "This script is about to push export branch $export"
+            _x git push
+            _release_step_write $to $step
+        fi
 
         from=$to
     done
-}
-
-release_copy() {
-    declare -r branches=$*
-    declare from=
-    declare to=
-    declare release=
-    declare export=
-
-    _check_no_changes
-
-    # steps 1, 2 and 3 from usage for release
+    
+    # steps 1, 2 from usage for release
     from=
+    first=
     for to in $branches
     do
-        if [[ -z "$from" ]]
-        then
-            # step 1 from usage for release
-            # install first branch
-            _switch $to
-            _prompt "You may need to install branch $to first in ANOTHER SESSION (using PATO GUI with POMs from $pwd/apex and $pwd/db)"
-        fi
+        for step in 1 2a 2b 2c
+        do
+            if [[ $step == "1" ]]
+            then
+                if [[ -n "$from" ]]; then continue; fi
+                release=
+            else
+                if [[ -z "$from" ]]; then continue; fi
+                release="release/$from-$to"                        
+            fi
+
+            if ! _release_step_skip $to $step
+            then
+                case $step in
+                    # install first branch
+                    1)  _switch $to
+                        _prompt "You may need to install (database only) for '$to' first in ANOTHER SESSION (using PATO GUI with POMs from $pwd/apex and $pwd/db)"
+                        ;;
+
+                    # copy to release branch
+                    2a) if [[ "$first" == "$to" ]]
+                        then
+                            copy $from $release
+                        else
+                            copy "export/$from" $release
+                        fi
+                        ;;
+
+                    # create a release export (only for the first, the others can use the previous export)
+                    2b) if [[ "$first" == "$to" ]]
+                        then
+                            _prompt "You must create an export (APEX and/or database) for '$to' in ANOTHER SESSION (using PATO GUI with POMs from $pwd/apex and $pwd/db)"
+                            _x git add .
+                            _x git commit -m "Created release $release" || true
+                        fi
+                        ;;
+
+                    # push the release branch
+                    2c) _prompt "This script is about to push branch $release"
+                        _x git push
+                        ;;
+                esac
+                _release_step_write $to $step
+            fi
+        done
         
+        from=$to
+        [[ -n "$first" ]] || first=$to
+    done
+
+    # step 3 from usage for release
+    from=
+    step=3
+    for to in $branches
+    do
         if [[ -n "$from" ]]
         then
-            # step 2 from usage for release
-            # step 2a
-            release="release/$from-$to"
-            copy $from $release
-            # step 2b
-            _prompt "Please create a release export (APEX and/or database) in ANOTHER SESSION (using PATO GUI with POMs from $pwd/apex and $pwd/db)"
-            # step 2c
-            _prompt "Pushing branch $release"
-            _x git push
+            release="release/$from-$to"                        
 
-            # step 3 from usage for release
-            # step 3a
-            _pull_request $release $to
-            # step 3b
-            _tag
-            url=$(git config --get remote.origin.url)
-            url=$(basename $url .git)
-            _prompt "Please ensure that the Pull Request from $release to $to has been accepted (go to $url)"
+            if ! _release_step_skip $to $step
+            then
+                _prompt "The contents of branch $to will be REPLACED (not MERGED) with the contents of $release"
+                _switch $release
+                # replace $to by $release but keep history
+                _x git merge --strategy=ours --no-commit $to
+                ! _x git commit -m"Replaced contents of branch $to by branch $release" || _x git push
+                # normal merge to $to
+                _switch $to
+                _x git merge $release
+                _prompt "You must fix any existing conflicts (using GitHub Desktop for instance)"
+                url=$(git config --get remote.origin.url)
+                url=$(basename $url .git)
+                # _switch $to
+                _tag $to
+                _x git push
+
+                _release_step_write $to $step
+            fi
         fi
         
         from=$to
     done
-}
-
-release_install() {
-    declare -r branches=$*
-    declare from=
-    declare to=
-
-    _check_no_changes
 
     # step 4 from usage for release
     from=
+    step=4
     for to in $branches
     do
         if [[ -n "$from" ]]
         then
-            # install all except the first branch
-            _switch $to
-            _prompt "Install branch $to first in ANOTHER SESSION (using PATO GUI with POMs from $pwd/apex and $pwd/db)"
+            if ! _release_step_skip $to $step
+            then
+                # install all except the first branch
+                _switch $to
+                _prompt "You must install branch $to first in ANOTHER SESSION (using PATO GUI with POMs from $pwd/apex and $pwd/db)"
+                _release_step_write $to $step
+            fi
         fi
         
         from=$to
     done
-}
 
-release() {
-    release_backup "$@"
-    release_copy "$@"
-    release_install "$@"
+    _release_state_file_cleanup # on success remove release state
 }
 
 # ----
@@ -538,7 +651,7 @@ case "$command" in
         test $# -ge 2 || usage 1
         $command $*
         ;;
-    release | release_backup | release_copy | release_install)
+    release)
         test $# -ge 2 || usage 1
         $command $*
         ;;